+++ ApolloState.m
@@ -18,7 +18,9 @@ NSInteger sReadPostMaxCount = 0;
 NSInteger sUnmuteCommentsVideos = 0; // 0=Default, 1=Remember from Full Screen, 2=Always
 
 BOOL sProxyImgurDDG = NO;
-BOOL sUseRedditNativeImageUpload = NO;
+NSInteger sImageUploadProvider = ImageUploadProviderImgur;
+
+NSString *sLatestRedditBearerToken = nil;
 
 BOOL sEnableBulkTranslation = NO;
 BOOL sAutoTranslateOnAppear = YES;